Core Modules
This commit adds the event bus and command queue. The event bus is responsible for handling and distributing event s to the event listeners. The command queue is to handle the commands in the core game logic. This commit also adds the global files for the core modules. It also adds the command templates but they aren't used currently.