UI Manager, UI Component template
This commit adds the UI manager. It is responsible for rendering and updating all the UI components. It also has a few smaller functionality including changing the z-index, enabling and hiding components, and passing events to components. This commit also adds the UI Component template which will be used for all future UI components.