Type checking and comments
This commit doesn't change any functionality. It adds comments and types to all of the components and code written in the previous commit. It also changes some type issues with other areas of the codebase. There are very minor actual code changes and any changes are to fix type errors.