Board Renderer
This commit finally adds some visuals to the game. You can see the game board, which is made up of many hexes. Each time you press play, every tile is assigned a random number and hex type to create the board. Some other small changes in this commit includes some new and modified configs, type fixes in largest_army and longest_road, and a new function in the Board component to get the hex type and number tile.