Board Renderer structure changes and comments.
This commit changes the structure of the board_renderer. It means the vertices and edges actually store there own pixel coordinates, so that it won't need to be calculated in the renderer. This is achieved by moving many of the functions to calculate coordinates into the Hex component, and when the board is created, it uses those functions to add the pixel coordinates to each vertex and board renderer. This commit also adds a few minor changes to types and comments in the board renderer and vertex / edge components.