Game UI - Part 1/2
This commit uses the components created in the last commit. It adds the frames for all the UI, all the buttons and text for the building UI. It also the functionality for buttons to actually do things. This commit fixes a load of bugs and small tweaks with the button component. This is part 1 out of 2 for the Game UI. The next commit will add the UI for the player panel and resource panel.