UI Scaling
This commit adds the ability for UI to scale based off of the screen size. It changes all the UI components to incorporate UI scaling, and changes the hex file to do the same. There is a new global file to store the scale factors. This commit also fixes a small bug with the Devcard button which was calling the wrong function.